[Feat] - useEffect 조건 변경
