+++ src/app/(nonAuthLayout)/guest/page.tsx
@@ -21,17 +21,59 @@ const Guest = () => {
         setReset();
     }, [setPortfolio, setReset]);
 
+    const onClickNextHandler = () => {
+        window.scrollTo({ top: 0, left: 0, behavior: "smooth" });
+        if (nav === "basicInfo") {
+            setNav("introduce");
+            return;
+        }
+        if (nav === "introduce") {
+            setNav("project");
+            return;
+        }
+        if (nav === "project") {
+            setNav("url");
+            return;
+        }
+    };
+
+    const onClickPrevHandler = () => {
+        window.scrollTo({ top: 0, left: 0, behavior: "smooth" });
+        if (nav === "introduce") {
+            setNav("basicInfo");
+            return;
+        }
+        if (nav === "project") {
+            setNav("introduce");
+            return;
+        }
+        if (nav === "url") {
+            setNav("project");
+            return;
+        }
+    };
+
     return (
-        <div className="flex justify-center bg-hihigray max-w-full min-h-full mx-auto sm:bg-white sm:flex-col sm:w-full">
-            <Navigation setNav={setNav} />
-            <div className="w-[800px] min-h-[750px] sm:w-full">
-                {nav === "basicInfo" && <UserInfo />}
-                {nav === "introduce" && <Introduction />}
-                {nav === "project" && <Project />}
-                {nav === "url" && <PortfolioUrl />}
+        <>
+            <div className="flex justify-center bg-hihigray max-w-full min-h-full mx-auto sm:bg-white sm:flex-col sm:w-full">
+                <Navigation setNav={setNav} />
+                <div className="w-[800px] min-h-[750px] sm:w-full">
+                    {nav === "basicInfo" && <UserInfo />}
+                    {nav === "introduce" && <Introduction />}
+                    {nav === "project" && <Project />}
+                    {nav === "url" && <PortfolioUrl />}
+                </div>
+                <Buttons />
             </div>
-            <Buttons />
-        </div>
+            <div className="flex justify-center gap-20 mt-4 text-primary">
+                <div className={`cursor-pointer ${nav === "basicInfo" && "opacity-0"}`} onClick={onClickPrevHandler}>
+                    PREV
+                </div>
+                <div className={`cursor-pointer ${nav === "url" && "opacity-0"}`} onClick={onClickNextHandler}>
+                    NEXT
+                </div>
+            </div>
+        </>
     );
 };
